third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/scoped-registry-append-does-not-upgrade-expected.txt

This is a testharness.js-based test.
[FAIL] Connecting a custom element candiate with null registry does not set the registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Connecting a custom element candiate with a scoped custom element registry does not change the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Inserting a custom element candiate with null registry does not change the registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Inserting a node from another document results in the custom element registry to be set and upgraded.
  assert_equals: expected null but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
